requirements_dev.txt
# contains all requirements to run the tests by running ```tox -e py36,py37```
tox>=3.5.2
pytest>=3.8.2
pytest-runner>=4.2
Click>=7.0
crowdai-api>=0.1.21
numpy>=1.16.2
recordtype>=1.3
matplotlib>=3.0.2
Pillow>=5.4.1
CairoSVG>=2.3.1
msgpack>=1.0.0
msgpack-numpy>=0.4.4.0
svgutils>=0.3.1
pyarrow>=0.13.0
importlib-metadata>=0.17
importlib-resources>=1.0.1
six>=1.12.0
timeout-decorator>=0.4.1
attrs
ushlex
gym==0.14.0
